Revert "Merge branch 'muon_dev' into falsetto"
This reverts commit 4f4dece53545687f2f6d88de175350f2f41712ed, reversing
changes made to acf194efee6d5c64a6855f0e9d7072f7150e1042.
