+++ basics/base_task.py
@@ -307,7 +307,7 @@ class BaseTask(pl.LightningModule):
         optimizer = build_object_from_class_name(
             optimizer_args['optimizer_cls'],
             torch.optim.Optimizer,
-            model if optimizer_args['optimizer_cls'] == 'modules.optimizer.muon.Muon_AdamW' else model.parameters(),
+            model.parameters(),
             **optimizer_args
         )
         return optimizer
