+++ configs/acoustic.yaml
@@ -72,12 +72,13 @@ sampling_steps: 20
 diff_accelerator: ddim
 diff_speedup: 10
 hidden_size: 256
-backbone_type: 'lynxnet2'
+backbone_type: 'lynxnet'
 backbone_args:
   num_channels: 1024
   num_layers: 6
   kernel_size: 31
   dropout_rate: 0.0
+  strong_cond: true
 main_loss_type: l2
 main_loss_log_norm: false
 schedule_type: 'linear'
@@ -106,25 +107,20 @@ lambda_aux_mel_loss: 0.2
 # train and eval
 num_sanity_val_steps: 1
 optimizer_args:
-  optimizer_cls: modules.optimizer.muon.Muon_AdamW
   lr: 0.0006
-  muon_args:
-    weight_decay: 0.1
-  adamw_args:
-    weight_decay: 0.0
 lr_scheduler_args:
-  step_size: 5000
-  gamma: 0.8
+  step_size: 10000
+  gamma: 0.75
 max_batch_frames: 50000
 max_batch_size: 64
 dataset_size_key: 'lengths'
 val_with_vocoder: true
 val_check_interval: 2000
 num_valid_plots: 10
-max_updates: 100000
+max_updates: 160000
 num_ckpt_keep: 5
-permanent_ckpt_start: 60000
-permanent_ckpt_interval: 10000
+permanent_ckpt_start: 80000
+permanent_ckpt_interval: 20000
 
 finetune_enabled: false
 finetune_ckpt_path: null
