+++ modules/backbones/__init__.py
@@ -1,13 +1,11 @@
 import torch.nn
 from modules.backbones.wavenet import WaveNet
 from modules.backbones.lynxnet import LYNXNet
-from modules.backbones.lynxnet2 import LYNXNet2
 from utils import filter_kwargs
 
 BACKBONES = {
     'wavenet': WaveNet,
-    'lynxnet': LYNXNet,
-    'lynxnet2': LYNXNet2,
+    'lynxnet': LYNXNet
 }
