+++ modules/fastspeech/tts_modules.py
@@ -62,7 +62,7 @@ class DurationPredictor(torch.nn.Module):
     """
 
     def __init__(self, in_dims, n_layers=2, n_chans=384, kernel_size=3,
-                 dropout_rate=0.1, offset=1.0, dur_loss_type='mse', arch='resnet'):
+                 dropout_rate=0.1, offset=1.0, dur_loss_type='mse'):
         """Initialize duration predictor module.
         Args:
             in_dims (int): Input dimension.
@@ -76,29 +76,16 @@ class DurationPredictor(torch.nn.Module):
         self.offset = offset
         self.conv = torch.nn.ModuleList()
         self.kernel_size = kernel_size
-        self.use_resnet = (arch == 'resnet')
         for idx in range(n_layers):
             in_chans = in_dims if idx == 0 else n_chans
-            if self.use_resnet:
-                self.conv.append(nn.Sequential(
-                    LayerNorm(in_chans, dim=1),
-                    nn.Conv1d(in_chans, n_chans, kernel_size, stride=1, padding=kernel_size // 2),
-                    nn.ReLU(),
-                    nn.Conv1d(n_chans, n_chans, 1),
-                    nn.Dropout(dropout_rate)
-                ))
-            else:
-                self.conv.append(nn.Sequential(
-                    nn.Identity(),  # this is a placeholder for ConstantPad1d which is now merged into Conv1d
-                    nn.Conv1d(in_chans, n_chans, kernel_size, stride=1, padding=kernel_size // 2),
-                    nn.ReLU(),
-                    LayerNorm(n_chans, dim=1),
-                    nn.Dropout(dropout_rate)
-                ))
-        if self.use_resnet and in_dims != n_chans:
-            self.res_conv = nn.Conv1d(in_dims, n_chans, 1)
-        else:
-            self.res_conv = None
+            self.conv.append(torch.nn.Sequential(
+                torch.nn.Identity(),  # this is a placeholder for ConstantPad1d which is now merged into Conv1d
+                torch.nn.Conv1d(in_chans, n_chans, kernel_size, stride=1, padding=kernel_size // 2),
+                torch.nn.ReLU(),
+                LayerNorm(n_chans, dim=1),
+                torch.nn.Dropout(dropout_rate)
+            ))
+
         self.loss_type = dur_loss_type
         if self.loss_type in ['mse', 'huber']:
             self.out_dims = 1
@@ -134,12 +121,8 @@ class DurationPredictor(torch.nn.Module):
         xs = xs.transpose(1, -1)  # (B, idim, Tmax)
         masks = 1 - x_masks.float()
         masks_ = masks[:, None, :]
-        for idx, f in enumerate(self.conv):
-            if self.use_resnet:
-                residual = self.res_conv(xs) if idx == 0 and self.res_conv is not None else xs
-                xs = residual + f(xs)
-            else:
-                xs = f(xs)
+        for f in self.conv:
+            xs = f(xs)  # (B, C, Tmax)
             if x_masks is not None:
                 xs = xs * masks_
         xs = self.linear(xs.transpose(1, -1))  # [B, T, C]
