+++ modules/fastspeech/variance_encoder.py
@@ -46,8 +46,7 @@ class FastSpeech2Variance(nn.Module):
                 dropout_rate=dur_hparams['dropout'],
                 kernel_size=dur_hparams['kernel_size'],
                 offset=dur_hparams['log_offset'],
-                dur_loss_type=dur_hparams['loss_type'],
-                arch=dur_hparams['arch']
+                dur_loss_type=dur_hparams['loss_type']
             )
 
     def forward(
